server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
Pillow==11.0.0
opencv-python==4.10.0.84
numpy==1.26.4
pydantic==2.10.3
pydantic-settings==2.7.0
ultralytics==8.3.65
matplotlib==3.10.0
scipy==1.14.1
scikit-learn==1.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
python-magic==0.4.27
pytest==8.3.4
pytest-asyncio==0.25.0